refactor: migrate app-level layouts to new PageLayout system [REC-001]
