+++ app/(auth)/layout.tsx
@@ -1,9 +1,8 @@
 'use client';
 
 import React from 'react';
-import { Box } from '@mui/material';
+import { PageLayout } from '@/components/layout/PageLayout';
 import Navigation from '@/components/Navigation';
-import PageLayout from '@/components/PageLayout';
 
 export default function AuthLayout({
     children,
@@ -11,30 +10,12 @@ export default function AuthLayout({
     children: React.ReactNode;
 }) {
     return (
-        <PageLayout>
-            <Box
-                sx={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    minHeight: '100vh',
-                    bgcolor: 'background.default',
-                }}
-            >
-                <Navigation />
-                <Box
-                    component="main"
-                    sx={{
-                        flex: 1,
-                        display: 'flex',
-                        flexDirection: 'column',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        py: 4,
-                    }}
-                >
-                    {children}
-                </Box>
-            </Box>
+        <PageLayout
+            layout="auth"
+            footer={{ show: false }}
+        >
+            <Navigation />
+            {children}
         </PageLayout>
     );
 }
